Move AssetsHelper
