+++ src/Asset/Conditions/FileTypeConditionRule.php
@@ -5,12 +5,13 @@ declare(strict_types=1);
 namespace CraftCms\Cms\Asset\Conditions;
 
 use craft\base\ElementInterface;
-use craft\helpers\Assets as AssetsHelper;
+use CraftCms\Cms\Asset\AssetsHelper;
 use CraftCms\Cms\Asset\Elements\Asset;
 use CraftCms\Cms\Condition\BaseMultiSelectConditionRule;
 use CraftCms\Cms\Element\Conditions\Contracts\ElementConditionRuleInterface;
 use CraftCms\Cms\Element\Queries\AssetQuery;
 use CraftCms\Cms\Element\Queries\Contracts\ElementQueryInterface;
+use Override;
 
 use function CraftCms\Cms\t;
 
@@ -26,7 +27,7 @@ class FileTypeConditionRule extends BaseMultiSelectConditionRule implements Elem
         return ['kind'];
     }
 
-    #[\Override]
+    #[Override]
     protected function operators(): array
     {
         return [
