+++ src/Asset/Validation/Rules/AssetLocationRule.php
@@ -5,8 +5,7 @@ declare(strict_types=1);
 namespace CraftCms\Cms\Asset\Validation\Rules;
 
 use Closure;
-use craft\helpers\Assets;
-use craft\helpers\Assets as AssetsHelper;
+use CraftCms\Cms\Asset\AssetsHelper;
 use CraftCms\Cms\Asset\Elements\Asset;
 use CraftCms\Cms\Cms;
 use CraftCms\Cms\Support\Facades\Assets as AssetsService;
@@ -38,7 +37,7 @@ readonly class AssetLocationRule implements ValidationRule
             return;
         }
 
-        [$folderId, $filename] = Assets::parseFileLocation($value);
+        [$folderId, $filename] = AssetsHelper::parseFileLocation($value);
 
         $hasNewFolderId = $folderId !== $this->asset->{$this->folderIdAttribute};
         $hasNewFilename = $filename !== $this->asset->{$this->filenameAttribute};
