+++ src/Shared/Enums/TimePeriod.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace CraftCms\Cms\Shared\Enums;
 
+use function CraftCms\Cms\t;
+
 /**
  * PeriodType defines time period units.
  */
@@ -16,4 +18,17 @@ enum TimePeriod: string
     case Weeks = 'weeks';
     case Months = 'months';
     case Years = 'years';
+
+    public function label(): string
+    {
+        return match ($this) {
+            TimePeriod::Seconds => t('Seconds'),
+            TimePeriod::Minutes => t('Minutes'),
+            TimePeriod::Hours => t('Hours'),
+            TimePeriod::Days => t('Days'),
+            TimePeriod::Weeks => t('Weeks'),
+            TimePeriod::Months => t('Months'),
+            TimePeriod::Years => t('Years'),
+        };
+    }
 }
