+++ src/User/Users.php
@@ -5,9 +5,9 @@ declare(strict_types=1);
 namespace CraftCms\Cms\User;
 
 use Craft;
-use craft\helpers\Assets as AssetsHelper;
 use craft\helpers\UrlHelper;
 use craft\web\Request;
+use CraftCms\Cms\Asset\AssetsHelper;
 use CraftCms\Cms\Asset\Data\Volume;
 use CraftCms\Cms\Asset\Elements\Asset;
 use CraftCms\Cms\Asset\Exceptions\ImageException;
