+++ yii2-adapter/src/Event/EventCompatibility.php
@@ -19,6 +19,7 @@ use craft\fields\Assets as LegacyAssetsField;
 use craft\fields\BaseOptionsField as LegacyBaseOptionsField;
 use craft\fields\Link as LegacyLinkField;
 use craft\fields\Matrix as LegacyMatrixField;
+use craft\helpers\Assets;
 use craft\imagetransforms\ImageTransformer;
 use craft\models\FieldLayout;
 use craft\services\Addresses;
@@ -90,6 +91,11 @@ readonly class EventCompatibility
         LegacyLinkField::registerEvents();
         LegacyMatrixField::registerEvents();
         FieldEvents::registerEvents();
+
+        /**
+         * Helpers
+         */
+        Assets::registerEvents();
 
         /**
          * Services
